[rs] Make `serde` optional
This commit makes the `serde` (and `hex`) dependencies optional, allowing to remove most transitive dependencies when serialization is not required. `serde` is enabled by default, despite the original issue proposing to make it opt-in. Serialization is used by most of the consumers, and it helps with enabling `hex` only when `serde` is also enabled.

Closes open-flash/swf-tree#73