[all] Replace `Unknown` tag with `Raw`
The `Raw` tag serves as a generalized catch-all tag for invalid tags. `Unknown` represented tags with a code that the parser did not know or that the emitter wrote as-is. The new `Raw` tag extends it to also support invalid tag bodies or incomplete tag headers. Emitters are expected to append its content as-is.

The main difference is that the `data` field now contains the whole tag, not only its body.